requirements.txt
# Core Data Science Libraries
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0

# Machine Learning & Deep Learning
tensorflow==2.13.0
xgboost==1.7.6

# Explainability
shap==0.42.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Web Framework
flask==2.3.2
werkzeug==2.3.6

# Utilities
joblib==1.3.1
pyyaml==6.0
python-dotenv==1.0.0

# Optional: OpenAI integration
openai==0.27.8

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Jupyter
jupyter==1.0.0
ipykernel==6.25.0

# Data Quality
imbalanced-learn==0.11.0
